Revert "Merge branch 'develop' of https://github.com/PaddlePaddle/ERNIE into develop"
This reverts commit e3cc7800163ebc9daa7002beaf377bb5f6cd4108, reversing
changes made to bb18727ba01a1666b474e922c860ce1467b242c9.

+++ ernie/moe/moe_all_gather_layer.py
@@ -392,9 +392,7 @@ class AlltoAllSmart(paddle.autograd.PyLayer):
 
                 recv_mask_alltoall_out = paddle.concat(recv_mask_alltoall_out, 0)
                 distributed_input_to_alltoall_out = paddle.maximum(
-                    (recv_mask_alltoall_out.cumsum() - 1).astype(
-                        recv_mask_alltoall_out.dtype
-                    ),
+                    recv_mask_alltoall_out.cumsum() - 1,
                     paddle.zeros([1], dtype=recv_mask_alltoall_out.dtype),
                 )
                 distributed_input_to_alltoall_out = (
